chore: 1/120th of the owl
