+++ prisma/schema.prisma
@@ -25,8 +25,9 @@ model User {
   address   Address?
 
   // auth
-  email        String  @unique
-  passwordHash String?
+  OAuthAccounts OAuthAccount[]
+  email         String         @unique
+  passwordHash  String?
 
   // permissions
   admin Boolean? @default(false)
@@ -37,6 +38,14 @@ model User {
 
   // sessions
   sessions Session[]
+}
+
+model OAuthAccount {
+  id             String @id @default(cuid())
+  providerId     String
+  providerUserId String
+  userId         String
+  user           User   @relation(fields: [userId], references: [id])
 }
 
 model Category {
